requirements.txt
# 金融研报助手 - 依赖列表
# 除 LLM API 外，100% 本地化

# PDF 解析
pdfplumber==0.11.0
PyMuPDF==1.24.0
pdf2image==1.16.3

# NLP 处理
hanlp==2.1.0
snownlp==0.12.3
jieba==0.42.1
spacy==3.7.0

# 向量模型（本地运行）
sentence-transformers==2.3.0

# 数据库
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Web 框架
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3

# 测试
pytest==8.0.0
